requirements.txt
Django==1.9.6
django-filter==0.13.0
djangorestframework==3.3.3
django-rest-swagger==0.3.6
djangorestframework-jwt==1.8
Markdown==2.6.6
psycopg2==2.6.1
Pillow==3.2.0
requests==2.10.0
# Requires libffi-dev and libssl-dev to be installed on the system
cryptography==1.3.2
facebook-sdk==1.0.0
